requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==4.0.1
django-heroku==0.3.1
django-widget-tweaks==1.4.12
gunicorn==20.1.0
Pillow==9.0.1
psycopg2==2.9.3
python-decouple==3.6
sqlparse==0.4.2
tzdata==2021.5
whitenoise==6.0.0
django-storages==1.12.3
boto3==1.21.8